+++ modules/evaluation-server/src/Domain/Shared/TestData.cs
@@ -1,6 +1,4 @@
-using Domain.Shared;
-
-namespace TestBase;
+namespace Domain.Shared;
 
 public static class TestData
 {
