+++ modules/evaluation-server/tests/Domain.UnitTests/Domain.UnitTests.csproj
@@ -14,6 +14,7 @@
 
     <ItemGroup>
         <ProjectReference Include="..\TestBase\TestBase.csproj" />
+        <ProjectReference Include="..\..\src\Domain\Domain.csproj" />
     </ItemGroup>
 
 </Project>
