+++ modules/evaluation-server/tests/Streaming.UnitTests/Connections/ConnectionBuilder.cs
@@ -1,4 +1,3 @@
-using TestBase;
 using System.Net.WebSockets;
 using Domain.Shared;
 using Moq;
