+++ modules/evaluation-server/tests/Streaming.UnitTests/Messages/RequestValidatorTests.cs
@@ -1,5 +1,4 @@
-﻿using TestBase;
-using System.Net.WebSockets;
+﻿using System.Net.WebSockets;
 using Domain.Shared;
 using Microsoft.Extensions.Internal;
 using Moq;
