+++ modules/evaluation-server/tests/TestBase/TestBase.csproj
@@ -19,8 +19,4 @@
         <PackageReference Include="Verify.Xunit" Version="19.6.0" />
     </ItemGroup>
 
-    <ItemGroup>
-        <ProjectReference Include="..\..\src\Domain\Domain.csproj" />
-    </ItemGroup>
-
 </Project>
